# openapi.yml
openapi: '3.0.3'
info:
  title: B2B Usage-Based Billing Platform API
  version: '1.0.0'
...
paths:
  /v1/auth/register:
    post:
      summary: Create a new account and send email verification
      operationId: register
      tags: [Auth]
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '202':
          description: Registration accepted. Verification email has been sent
          content:
            'application/json':
              schema:
                type: object
        '400':
          description: | 
            Bad request 
            
            Additional codes:
            - 2001: Required name
            - 2002: Required email
            - 2003: Required password
            - 2004: Invalid name
            - 2005: Invalid email
            - 2006: Invalid password
            - 2007: Email is already in use
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/BadRequestResponse'
        '500':
          description: Internal error
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/InternalErrorResponse'
  /v1/auth/email/verify:
    post:
      summary: Verify email
      operationId: verifyEmail
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
      responses:
        '204':
          description: Email verified
        '400':
          description: | 
            Bad request 
            
            Additional codes:
            - 2001: Required token
            - 2002: Invalid token
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/BadRequestResponse'
        '500':
          description: Internal error
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/InternalErrorResponse'
  /v1/auth/email/resend-verification:
    post:
      summary: Resend email verification
      operationId: resendEmailVerification
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResendEmailVerificationRequest'
      responses:
        '204':
          description: Email verification has been resent
        '400':
          description: |
            Bad request 
            
            Additional codes:
            - 2001: Required email
            - 2002: Invalid email
            - 2003: Email is already verified
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/BadRequestResponse'
        '500':
          description: Internal error
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/InternalErrorResponse'
  /v1/auth/login:
    post:
      summary: Create a new session
      operationId: login
      tags: [Auth]
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Session created
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /v1/auth/refresh:
    post:
      summary: Refresh an access token of session
      operationId: refreshSession
      tags: [Auth]
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/RefreshSessionRequest'
      responses:
        '200':
          description: Session refreshed
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/AuthResponse'
...
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    BadRequestResponse:
      type: object
      properties:
        message:
          type: string
          example: Invalid request format
        code:
          type: integer
          example: 1001
    InternalErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: Internal error
    RegisterRequest:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          description: |
            Rules:
            - required field
            - length between 3 and 100
            - first character is letter
        email:
          type: string
          format: email
          description: |
            Rules:
            - required field
            - length no more than 100
            - email format
        password:
          type: string
          format: password
          description: |
            Rules:
            - required field
            - length between 8 and 64
            - at least one letter
            - at least one number
    VerifyEmailRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          format: uuid
    ResendEmailVerificationRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          description: |
            Rules:
            - required field
            - length no more than 100
            - email format
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    RefreshSessionRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
    AuthResponse:
      type: object
      required:
        - access_token
        - refresh_token
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
...